refactor: organize and clean up project file structure
