+++ lib/modules/bbun/presentation/pages/detail_page.dart
@@ -1,6 +1,6 @@
 import 'package:auto_route/auto_route.dart';
-import 'package:bbun/presentation/widgets/bbun_notebook.dart';
-import 'package:bbun/presentation/widgets/bbun_pressable.dart';
+import 'package:bbun/modules/bbun/presentation/widgets/bbun_notebook.dart';
+import 'package:bbun/modules/bbun/presentation/widgets/bbun_pressable.dart';
 import 'package:bbun/routes/app_router.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_svg/flutter_svg.dart';
