+++ lib/modules/bbun/presentation/pages/main_page.dart
@@ -5,8 +5,10 @@ import 'package:bbun/di/locator.dart';
 import 'package:bbun/gen/assets.gen.dart';
 import 'package:bbun/modules/bbun/presentation/bloc/bbun_bloc.dart';
 import 'package:bbun/modules/bbun/presentation/bloc/bbun_list_bloc.dart';
+import 'package:bbun/modules/bbun/presentation/pages/profile_edit_page.dart';
 import 'package:bbun/modules/bbun/presentation/widgets/bbun_pressable.dart';
 import 'package:bbun/modules/user/presentation/bloc/auth_bloc.dart';
+import 'package:bbun/routes/app_router.dart';
 import 'package:flutter/material.dart';
 import 'package:bbun/modules/bbun/presentation/widgets/bbun_card.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
@@ -288,8 +290,8 @@ class _MainPageState extends State<MainPage>
                                   width: 317 * scale,
                                   child: BbunPressable(
                                     onPressed: () {
-                                      print(bbunListState.getTotal);
-                                      print(context.read<AuthBloc>().state);
+                                      context.router
+                                          .push(const ProfileEditRoute());
                                     },
                                     decoration: BoxDecoration(
                                       color: Color(0xFFFFE24A),
