+++ lib/routes/app_router.dart
@@ -1,7 +1,8 @@
 import 'package:auto_route/auto_route.dart';
 import 'package:bbun/modules/bbun/presentation/pages/onboarding_page.dart';
 import 'package:bbun/modules/bbun/presentation/pages/main_page.dart';
-import 'package:bbun/presentation/pages/detail_page.dart';
+import 'package:bbun/modules/bbun/presentation/pages/detail_page.dart';
+import 'package:bbun/modules/bbun/presentation/pages/profile_edit_page.dart';
 import 'package:bbun/routes/auth_guard.dart';
 
 part 'app_router.gr.dart';
